// web/types/loading-messages.ts
export const enum LOADING_TEXT {
	// initial state
	EMPTY = "",

	// NFT
	UPLOAD_NFT = "Uploading NFT metadata to IPFS...",
	MINT_NFT = "Minting NFT...",
	LIST_NFT_FOR_SALE = "Listing the NFT for sale...",
	LIST_NFT = "Listing NFT...",
	BUY_NFT = "Buying NFT...",

	// Others
	UPLOAD_IMAGE = "Uploading image to IPFS: ",
	WEB3 = "Getting Web3 Provider...",
	WAIT_TRANSACTION = "Waiting for transaction to complete...",
	LOAD = "Loading..."
}